split Merge() test
