+++ bitcask_test.go
@@ -1036,70 +1036,6 @@ func TestMaxDatafileSize(t *testing.T) {
 	})
 }
 
-func TestMerge(t *testing.T) {
-	var (
-		db  *Bitcask
-		err error
-	)
-
-	assert := assert.New(t)
-
-	testdir, err := ioutil.TempDir("", "bitcask")
-	assert.NoError(err)
-
-	t.Run("Setup", func(t *testing.T) {
-		t.Run("Open", func(t *testing.T) {
-			db, err = Open(testdir, WithMaxDatafileSize(32))
-			assert.NoError(err)
-		})
-
-		t.Run("PutBytes", func(t *testing.T) {
-			err := db.PutBytes([]byte("foo"), []byte("bar"))
-			assert.NoError(err)
-		})
-
-		s1, err := db.Stats()
-		assert.NoError(err)
-		assert.Equal(0, s1.Datafiles)
-		assert.Equal(1, s1.Keys)
-
-		t.Run("PutBytes", func(t *testing.T) {
-			for i := 0; i < 10; i++ {
-				err := db.PutBytes([]byte("foo"), []byte("bar"))
-				assert.NoError(err)
-			}
-		})
-
-		s2, err := db.Stats()
-		assert.NoError(err)
-		assert.Equal(5, s2.Datafiles)
-		assert.Equal(1, s2.Keys)
-		assert.True(s2.Size > s1.Size)
-
-		t.Run("Merge", func(t *testing.T) {
-			err := db.Merge()
-			assert.NoError(err)
-		})
-
-		s3, err := db.Stats()
-		assert.NoError(err)
-		assert.Equal(2, s3.Datafiles)
-		assert.Equal(1, s3.Keys)
-		assert.True(s3.Size > s1.Size)
-		assert.True(s3.Size < s2.Size)
-
-		t.Run("Sync", func(t *testing.T) {
-			err = db.Sync()
-			assert.NoError(err)
-		})
-
-		t.Run("Close", func(t *testing.T) {
-			err = db.Close()
-			assert.NoError(err)
-		})
-	})
-}
-
 func TestGetErrors(t *testing.T) {
 	//assert := assert.New(t)
 
@@ -1389,69 +1325,6 @@ func TestDeleteErrors(t *testing.T) {
 
 		err = db.Delete([]byte("foo"))
 		assert.Error(err)
-	})
-	*/
-}
-
-func TestMergeErrors(t *testing.T) {
-	assert := assert.New(t)
-
-	t.Run("RemoveDatabaseDirectory", func(t *testing.T) {
-		testdir, err := ioutil.TempDir("", "bitcask")
-		assert.NoError(err)
-		defer os.RemoveAll(testdir)
-
-		db, err := Open(testdir, WithMaxDatafileSize(32))
-		assert.NoError(err)
-
-		assert.NoError(os.RemoveAll(testdir))
-
-		err = db.Merge()
-		assert.Error(err)
-	})
-
-	/* todo
-	t.Run("EmptyCloseError", func(t *testing.T) {
-		testdir, err := ioutil.TempDir("", "bitcask")
-		assert.NoError(err)
-		defer os.RemoveAll(testdir)
-
-		db, err := Open(testdir)
-		assert.NoError(err)
-
-		mockDatafile := new(mockDatafile)
-		mockDatafile.On("Close").Return(ErrMockError)
-		db.curr = mockDatafile
-
-		err = db.Merge()
-		assert.Error(err)
-		assert.Equal(ErrMockError, err)
-	})
-	*/
-
-	/* todo
-	t.Run("ReadError", func(t *testing.T) {
-		testdir, err := ioutil.TempDir("", "bitcask")
-		assert.NoError(err)
-		defer os.RemoveAll(testdir)
-
-		db, err := Open(testdir, WithMaxDatafileSize(22))
-		assert.NoError(err)
-
-		assert.NoError(db.PutBytes([]byte("foo"), []byte("bar")))
-		assert.NoError(db.PutBytes([]byte("bar"), []byte("baz")))
-
-		mockDatafile := new(mockDatafile)
-		mockDatafile.On("Close").Return(nil)
-		mockDatafile.On("ReadAt", int64(0), int64(30)).Return(
-			Entry{},
-			ErrMockError,
-		)
-		db.datafiles[0] = mockDatafile
-
-		err = db.Merge()
-		assert.Error(err)
-		assert.Equal(ErrMockError, err)
 	})
 	*/
 }
@@ -1958,27 +1831,6 @@ func TestLocking(t *testing.T) {
 	assert.NoError(err)
 	defer db.Close()
 
-	_, err = Open(testdir)
-	assert.Error(err)
-}
-
-func TestLockingAfterMerge(t *testing.T) {
-	assert := assert.New(t)
-
-	testdir, err := ioutil.TempDir("", "bitcask")
-	assert.NoError(err)
-
-	db, err := Open(testdir)
-	assert.NoError(err)
-	defer db.Close()
-
-	_, err = Open(testdir)
-	assert.Error(err)
-
-	err = db.Merge()
-	assert.NoError(err)
-
-	// This should still error.
 	_, err = Open(testdir)
 	assert.Error(err)
 }
